Change heuristic for unmodified upstream drs to LLVM SVN revisions
The new heuristic is much faster, getting throughput up to around 2
commits per second.  Translation is still too slow, but it's somewhere
around 5-10x faster than it was before.

The old heuristic was to compare the commit timestamp of the last change
to a directory, using `git log -- <dir>/`.  This was very slow because
it requires a Git graph traversal each time we need to comopare.

The new heuristic is to track which LLVM SVN revision is the "base" for
a commit, and take the content from the higher rev number.

- This is exactly correct for content untouched from LLVM upstream.
    - Unless we somehow merge an llvm.org release branch and llvm.org
      master together, and the release branch rev number is bigger.
- This might be a reasonable heuristic for content introduced in a
  downstream branch that is upstream of the current one.  I.e., assume
  that the commit that has merged from llvm.org most recently has the
  "correct" content.
    - "Reasonable" is probably good enough.  If a downstream branch
      wants precision it should add a `repeat` declaration for some
      upstream branch whose content it wants to match.
- There is a known bug when a root is unrelated to LLVM upstream.  The
  commit will be rejected.  The fix is to treat that as "r0", but until
  we have tests in place that's a pretty scary catch-all.