Fix a read-after-write bug in split2monodb
The file_stream abstraction had some logic around the size of the file
in seek_to_read that only makes sense for mmap'ed files whose content
doesn't change.  For read/write streams, we should (mostly) ignore the
initial number of bytes in the file.

This also renames a bunch of fields and accessors to make it clear that
they will be stale after writing.