Handle already-translated commits
Handle a quirk of our internal history, where there was once another git-svn
mirror that had slightly different history from the canonical ones at
llvm.org.  Solution is to check if the commit in a queue is already
mapped to a monorepo commit (which git-svn commits all will be, through
their git-svn-id:) and skip it if so.