Fix eager computation of boundary parent revs
Replace `&& rev` with `|| !rev` when deciding whether to try
to compute a monorepo rev or note it from split commit.  This
fixes a bug when interleaving a second-level branch, which is
downstream of another downstream branch, or when updating a
pre-existing branch, since a fresh branch layered directly
on top of SVN will parse the rev from the commit message.