Give the first parent priority for tracked dirs (for real)
Tracked directories should come from the first parent even when not
building a first-parent commit.  Otherwise, tracked dirs will swing
somewhat arbitrarily in side histories.

Note that this change does *not* fix the same problem in second-level
downstream branches that track fewer directories.  Added a FIXME for
that, although it's not clear exactly how to fix it.

Note that construct_tree might be easier to follow if the big loop is
split in two; a first pass to figure out what we want to do and a second
pass to actually do it.