Heal repos in mt-configs
Become robust to repository configuration changes by healing clone
configuration `mt_setup`.  I refactored the clone/update code into
functions to make it easier to follow the logic.  Note that I also
deleted the sleep logic since it wasn't working anyway (it was checking
`[ -d $clone ]` before setting `clone`) and it seems like we didn't need
it.

This is a follow-up to one of the previous commits which required cleaning the
workspace in CI.  Any change to the repositories in the config could
potentially have the same impact, so healing like this is safer.