mt-generate: Fix a double-negative in repeated dirs
There were two ways to add `--not X` to the command-line in
`mt_list_commits`, causing one of them to be double-negative.  The
double-not caused the commits already generated on a branch to repeated
again recursively, instead of the intended effect of avoiding repeating
the same commit twice when `mt-repeat` is insufficient.

This commit fixes the double-negative by collecting all the `--not`s
together and only specifying `--not` once.

Here's a more complete description of the old behaviour.  If the source
branch A for repeated commits had a commit X (subject "X") to repeat
onto branch B (whose head starts as Y), the first run would transform:

    A: X
    B: Y

into:

    A: X --\
            \
    B: Y --- X'

where `X'` has the subject "Merge: X".  This is correct.  However, if
run again with no changes, it would create:

    A: X --\
            \  ,---,
    B: Y --- X' --- X''

instead of doing nothing, where `X''` has the subject "Merge: Merge: X".
Running a third time would create `X'''` with a more deeply-nested
subject.

Now, the second run does nothing.