mt-config: Generate apple/stable/20160127 from */swift-3.0-branch
apple/stable/20160127 is the first of a series of historical branches
that feed into the current "stable" branches on the LLVM repo.  Add
configuration to generate it.

I found which historical branches are interesting by running:

```
$ git apple-llvm count-merged --uniq --reverse --min=1 \
    github/apple/stable --not llvm.org/master |
  grep -v 'Merge pull request'
```

on an LLVM clone and then digging deeper, replacing
"github/apple/stable" with the second parent of interesting merge
points.  I documented the results in apple.mt-config.  Except for
swift-3.0-branch, I have only checked the 'llvm' repo so far.  It's not
clear if 'clang', etc., will have an identical set of interesting
points.