mt-config: Generate remaining apple/stable/* branches
Generate stable branches corresponding to swift-5.0-branch and
swift-5.1-branch, excluding LLDB (since LLDB depends on the 'swift'
repo).  There are three here:

- apple/stable/20180719, the original branch point for swift-5.0-branch.
- apple/stable/20180912, the rebranch from swift/upstream-with-swift.
- apple/stable/20190104, the branch for swift-5.1-branch.

This gives us proper modelling for most of the dependencies in the
history of the current "stable" branch.  Note that clang-tools-extra and
libcxx had some history in the latter two branches.