mt-config: Add branches to incorporate LLDB
LLDB's branching strategy is completely different from the other LLVM
sub-projects, so the history this generates is not ideal.  We can
iterate and regenerate if this is not good enough.

For swift/swift-4.1-branch, LLDB needs the new `no-pass` feature in
`repeat` to follow the second-parent history before the branch was
fixed.

As a drive-by, also:

- Take advantage of new archival branches on the split repos for
  archive/original/swift-5.0-branch, rather than uses Git hashes
  directly for the original branch points.
- Fix the `apple/stable/*` date for swift-5.0-branch, changing it from
  when the branch was recreated (20180912) to the commit date of the
  branch point (20180801).