split2mono: Prevent 8bf1494af87f from being identified as r354826
8bf1494af87f is a split commit in github.com/apple/swift-clang that is a
strange style of cherry-pick of r354826.  The strange bit is that the
committer and author date match.  I thought this wouldn't occur, since
when using `git-cherry-pick` a collision should only happen when the
following all happen in the same second:

- Push to SVN.
- Wait for `git-svn` mirrors to update.
- Fetch.
- Cherry-pick to a branch.

This being vanishingly unlikely, I had assumed only a malicious actor
would add such a commit, and based the heuristics for "is this a
cherry-pick or a real llvm.org commit?" on seeing a mismatch.

Unfortunately there's at least one exception, the above cherry-pick.
Somehow, the committer and author metadata (including name and author,
which we're not checking yet) match each other exactly.  The
cherry-picker must have done something manual here that accidentally
wiped the author information, since they don't match the upstream
commit.

It's possible there are other cases of this in our histories.  It should
be possible to detect with a scan, looking for `git-svn-id:` where the
metadata matches perfectly but the authorship does not match the
upstream commit.

For now, just blacklist that commit.  We'll need to regenerate the
monorepo to avoid it (we can do that now, or sometime later in a v2).

Follow up commits will strengthen the existing heuristic to check name
and email and sniff for `(cherry picked ...)`, but none of those would
have rejected this particular commit.