split2mono: Also check for "(cherry picked ...)" when looking up revs
Further strengthen the heuristic for whether a `git-svn-id:`-tagged
commit was cherry-picked by seeing if the next line starts with
`(cherry`.  It's exceedingly unlikely that the author/commit metadata
will match exactly AND someone used `git cherry-pick -x`, but since this
is easy to check for we may as well.