mt-generate: Canonicalize 'MERGED_MT_CONFIG' before 'cd'
Canonicalize `MERGED_MT_CONFIG` before changing directory to `GIT_DIR`.
When there's an upstream this is a no-op since it will already be an
absolute path to a file generated by `mktemp`, but without an upstream
it will be the same as `MT_CONFIG`.

Also fix trailing whitespace in test/mt/generate/splitrefs.test as a
drive-by.