split2mono: Strengthen heuristic for detecting upstream git-svn commits
Improve the heuristic for detecting upstream git-svn commits.  Now we
only parse the commit metadata looking for `git-svn-id:` when (a) the
commit is not a merge, (b) the parent commit (if any) has not been
generated by split2mono, and (c) the author and committer info match
exactly.  Before this commit, the only check was (c).

This required some (perhaps overdue) refactoring of compute_rev to make
the logic for split commits and monorepo commits clear (and, in some
cases, different).  A few functions in `git_cache` were renamed for
clarity.

The tests use an invented history that looks like an LLVM subproject,
with `git-svn-id:` for the split commits and `llvm-rev:` for the
monorepo commits.