mt-generate: Ensure that `repeat` is a no-op when it should be
`mt-translate-ref` is sometimes (but rarely) called when there are new
commits on the branch getting repeated, but none of them affect the
"repeated" dirs.  Ensure that we don't generate merge commits for those,
since it makes the history extra confusing (even without this change,
the content was correct).

As a drive-by, avoid comparing trees twice.