apple-v5: Repeat `dir`s in favour of `repeat`s
The `split2mono` tool supports "auto" repeats of `dir`s that are shared
between branches.  When we can use that it's more precise than a
`repeat`, because it will follow `--ancestry-path` instead of relying on
`--first-parent` (which could fail to generate a good history).  Prefer
that where possible.

Also fixed two things as drive-bys:

- Filled in the start directive for swift/master/start, which had been
  half typed in before.
- Fixed spelling for the clang-tools-extra `dir` on
  apple/master/root-start.