Make sure to export GIT_DIR in git-apple-llvm-fwd
This was setting GIT_DIR, but since it wasn't exported the git
commands would just create a git repo in the current directory
instead.

Now that we're exporting this, we can also clean up the place where we
would `cd` in to the directory. This was rather interestingly causing
us to use the expected git dir on the second and subsequent runs.