Pin gitdb2 to version 2.0.6
We have requirement for GitPython==2.1.11 which in turn requires GitDB.
There’s been a new version of GitDB and it has removed
`gitdb.utils.compat`. Since we pin GitPython to an older version, we’ll
have to pin GitDB to a corresponding working version too. Choosing 2.0.6
as that appears to be the last known working version.