fwd: Allow forwarding master as-is
Detecting whether there is anything to forward requires cleaning
out `refs/heads/*` and `refs/tags/*` from the local Git repo. This
uses a funky `git push` operation to allow globbed deletion, but
that means that Git is going to try to protect `refs/head/master`
from getting deleted.  Add a configuration to the local repo to
allow it to be deleted, since we don't actually care about it.